Client: DataBuffer class updated - use size_t instead of uint16_t

+++ Client/Src/DataBuffer.cpp
@@ -11,17 +11,17 @@ DataBuffer::DataBuffer(size_t _size):
 
 Handle DataBuffer::GetHead()
 {
-    Handle result={head,static_cast<uint16_t>(head<tail?tail-head-1:end-head-(tail>begin?0:1))};
+    Handle result={head,static_cast<size_t>(head<tail?tail-head-1:end-head-(tail>begin?0:1))};
     return result;
 }
 
 Handle DataBuffer::GetTail()
 {
-    Handle result={tail,static_cast<uint16_t>((head>=tail)?head-tail:end-tail)};
+    Handle result={tail,static_cast<size_t>((head>=tail)?head-tail:end-tail)};
     return result;
 }
 
-void DataBuffer::Commit(const Handle& handle, uint16_t usedSz)
+void DataBuffer::Commit(const Handle& handle, size_t usedSz)
 {
     uint8_t ** const target=handle.buffer==head?&head:&tail;
     *target+=usedSz;
@@ -29,9 +29,9 @@ void DataBuffer::Commit(const Handle& handle, uint16_t usedSz)
         *target=begin;
 }
 
-uint16_t DataBuffer::UsedSize()
+size_t DataBuffer::UsedSize()
 {
-    return static_cast<uint16_t>(head>=tail?head-tail:(end-tail)+(head-begin));
+    return static_cast<size_t>(head>=tail?head-tail:(end-tail)+(head-begin));
 }
 
 bool DataBuffer::IsHalfUsed()
